- Fixed attack speed of everything  - Changed all attackSpeed, dmg and life variables to FLOAT  - Implemented attacksPerSecond mechanic  - Removed unnecesary player initialization in the GameEngine constructor  - Added new bulletMechanic to test stuff out  - Created function to center bullets

+++ core/src/com/mygdx/game/Entities/Enemies/SatelliteOrbitEnemy.java
@@ -124,7 +124,7 @@ public class SatelliteOrbitEnemy extends Enemy {
 
     //SOBRESCRITO PARA QUE SEA INMORTAL
     @Override
-    public void decreaseLives(int lives) {
+    public void decreaseLives(float lives) {
 
     }
